magento/adobe-stock-integration#940: Refactoring
