+++ AdobeStockImageAdminUi/Ui/Component/Listing/Columns/UnlicensedOverlay.php
@@ -5,7 +5,7 @@
  */
 declare(strict_types=1);
 
-namespace Magento\MediaGalleryUi\Ui\Component\Listing\Columns;
+namespace Magento\AdobeStockImageAdminUi\Ui\Component\Listing\Columns;
 
 use Magento\Framework\View\Element\UiComponent\ContextInterface;
 use Magento\Framework\View\Element\UiComponentFactory;
