+++ AdobeStockImageAdminUi/etc/adminhtml/di.xml
@@ -9,9 +9,9 @@
     <type name="Magento\Cms\Block\Adminhtml\Wysiwyg\Images\Content">
         <plugin name="add_search_button" type="Magento\AdobeStockImageAdminUi\Plugin\AddSearchButton"/>
     </type>
-    <type name="Magento\MediaGalleryUi\Ui\Component\Listing\Columns\ConstructSourceItemUrl">
+    <type name="Magento\MediaGalleryUi\Ui\Component\Listing\Columns\SourceIconProvider">
         <arguments>
-            <argument name="imageSource" xsi:type="array">
+            <argument name="sourceIcons" xsi:type="array">
                 <item name="Adobe Stock" xsi:type="string">Magento_MediaGalleryUi::images/Astock.png</item>
             </argument>
         </arguments>
