+++ AdobeStockImageAdminUi/view/adminhtml/ui_component/media_gallery_listing.xml
@@ -28,4 +28,13 @@
             </filterSelect>
         </filters>
     </listingToolbar>
+    <columns name="media_gallery_columns">
+        <column sortOrder="1" name="overlay" component="Magento_Ui/js/grid/columns/overlay" class="Magento\AdobeStockImageAdminUi\Ui\Component\Listing\Columns\UnlicensedOverlay">
+            <settings>
+                <label translate="true">Overlay</label>
+                <visible>true</visible>
+                <sortable>false</sortable>
+            </settings>
+        </column>
+    </columns>
 </listing>
