+++ MediaGalleryUi/Ui/Component/Listing/Columns/SourceIconProvider.php
@@ -7,8 +7,11 @@ declare(strict_types=1);
 
 namespace Magento\MediaGalleryUi\Ui\Component\Listing\Columns;
 
+use Magento\Framework\App\Config\ScopeConfigInterface;
+use Magento\Framework\View\Asset\Repository as AssetRepository;
 use Magento\Framework\View\Element\UiComponent\ContextInterface;
 use Magento\Framework\View\Element\UiComponentFactory;
+use Magento\Store\Model\Store;
 use Magento\Ui\Component\Listing\Columns\Column;
 
 /**
@@ -16,26 +19,43 @@ use Magento\Ui\Component\Listing\Columns\Column;
  */
 class SourceIconProvider extends Column
 {
-    private $constructSourceItemUrl;
+    /**
+     * @var array
+     */
+    private $sourceIcons;
 
     /**
-     * SourceIconProvider constructor.
-     *
-     * @param ConstructSourceItemUrl $constructSourceItemUrl
+     * @var AssetRepository
+     */
+    private $assetRepository;
+
+    /**
+     * @var ScopeConfigInterface
+     */
+    private $scopeConfig;
+
+    /**
      * @param ContextInterface $context
      * @param UiComponentFactory $uiComponentFactory
+     * @param AssetRepository $assetRepository
+     * @param ScopeConfigInterface $scopeConfig
      * @param array $components
      * @param array $data
+     * @param array $sourceIcons
      */
     public function __construct(
         ContextInterface $context,
         UiComponentFactory $uiComponentFactory,
-        ConstructSourceItemUrl $constructSourceItemUrl,
+        AssetRepository $assetRepository,
+        ScopeConfigInterface $scopeConfig,
         array $components = [],
-        array $data = []
+        array $data = [],
+        array $sourceIcons = []
     ) {
         parent::__construct($context, $uiComponentFactory, $components, $data);
-        $this->constructSourceItemUrl = $constructSourceItemUrl;
+        $this->assetRepository = $assetRepository;
+        $this->scopeConfig = $scopeConfig;
+        $this->sourceIcons = $sourceIcons;
     }
 
     /**
@@ -48,11 +68,39 @@ class SourceIconProvider extends Column
     {
         if (isset($dataSource['data']['items']) && is_iterable($dataSource['data']['items'])) {
             foreach ($dataSource['data']['items'] as &$item) {
-                $item['source_icon_url'] =
-                    isset($item['source']) ? $this->constructSourceItemUrl->execute($item['source']) : null;
+                $item[$this->getData('name')] = $item[$this->getData('name')]
+                    ? $this->getSourceIconUrl($item[$this->getData('name')])
+                    : null;
             }
         }
 
         return $dataSource;
+    }
+
+    /**
+     * Construct source icon url based on the source code matching
+     *
+     * @param string $sourceName
+     *
+     * @return string|null
+     */
+    public function getSourceIconUrl(string $sourceName): ?string
+    {
+        return isset($this->sourceIcons[$sourceName])
+            ? $this->assetRepository->getUrlWithParams(
+                $this->sourceIcons[$sourceName],
+                ['_secure' => $this->getIsSecure()]
+            )
+            : null;
+    }
+
+    /**
+     * Check if store use secure connection
+     *
+     * @return bool
+     */
+    private function getIsSecure(): bool
+    {
+        return $this->scopeConfig->isSetFlag(Store::XML_PATH_SECURE_IN_ADMINHTML);
     }
 }
